+++ docs/CHANGELOG.md
@@ -1,1 +1,8 @@
+## [v1.1.0] - 2023-03-17
+ ### `accura` 
+ - Change barcode sequence to oligo ID in report and count matrix (Not performed in customized chemistry)
+ - Add UMI cutoff (default 500) to report display
+ - Add --skip_umi_correct to skip umi correction in count step for a large data set
+
+
 ## [released] - 2021-08-23